updated bug from py2
In Python 2, map returns a list, but in Python 3, it returns an iterator. When you wrap the map call with list(), you're effectively consuming the iterator and storing all its results in a list. Consequently, the map_generator variable becomes a list, not an iterator, which is why you encounter the error when trying to use next() on it.